+++ lib/active_record/connection_adapters/clickhouse_adapter.rb
@@ -3,6 +3,7 @@
 require 'clickhouse-activerecord/arel/visitors/to_sql'
 require 'clickhouse-activerecord/arel/table'
 require 'clickhouse-activerecord/migration'
+require 'active_record/connection_adapters/clickhouse/oid/array'
 require 'active_record/connection_adapters/clickhouse/oid/date'
 require 'active_record/connection_adapters/clickhouse/oid/date_time'
 require 'active_record/connection_adapters/clickhouse/oid/big_integer'
@@ -31,6 +32,7 @@ module ActiveRecord
             sslca: config[:sslca],
             read_timeout: config[:read_timeout],
             write_timeout: config[:write_timeout],
+            keep_alive_timeout: config[:keep_alive_timeout]
           }
         end
 
@@ -45,20 +47,18 @@ module ActiveRecord
     end
   end
 
-  class Relation
-
-    # Replace for only ClickhouseAdapter
+  module ClickhouseRelationReverseOrder
     def reverse_order!
-      orders = order_values.uniq
-      orders.reject!(&:blank?)
-      if self.connection.is_a?(ConnectionAdapters::ClickhouseAdapter) && orders.empty? && !primary_key
-        self.order_values = %w(date created_at).select {|c| column_names.include?(c) }.map{|c| arel_attribute(c).desc }
-      else
-        self.order_values = reverse_sql_order(orders)
-      end
+      return super unless connection.is_a?(ConnectionAdapters::ClickhouseAdapter)
+
+      orders = order_values.uniq.compact_blank
+      return super unless orders.empty? && !primary_key
+
+      self.order_values = %w(date created_at).select {|c| column_names.include?(c) }.map{|c| arel_attribute(c).desc }
       self
     end
   end
+  Relation.prepend(ClickhouseRelationReverseOrder)
 
   module TypeCaster
     class Map
@@ -85,9 +85,8 @@ module ActiveRecord
       def arel_table # :nodoc:
         @arel_table ||= ClickhouseActiverecord::Arel::Table.new(table_name, type_caster: type_caster)
       end
-
     end
-   end
+  end
 
   module ConnectionAdapters
     class ClickhouseColumn < Column
@@ -103,8 +102,13 @@ module ActiveRecord
         float: { name: 'Float32' },
         decimal: { name: 'Decimal' },
         datetime: { name: 'DateTime' },
+        datetime64: { name: 'DateTime64' },
         date: { name: 'Date' },
         boolean: { name: 'UInt8' },
+        uuid: { name: 'UUID' },
+
+        enum8: { name: 'Enum8' },
+        enum16: { name: 'Enum16' },
 
         int8:  { name: 'Int8' },
         int16: { name: 'Int16' },
@@ -181,6 +185,20 @@ module ActiveRecord
         end
       end
 
+      # `extract_scale` and `extract_precision` are the same as in the Rails abstract base class,
+      # except this permits a space after the comma
+
+      def extract_scale(sql_type)
+        case sql_type
+        when /\((\d+)\)/ then 0
+        when /\((\d+)(,\s?(\d+))\)/ then $3.to_i
+        end
+      end
+
+      def extract_precision(sql_type)
+        $1.to_i if sql_type =~ /\((\d+)(,\s?\d+)?\)/
+      end
+
       def initialize_type_map(m) # :nodoc:
         super
         register_class_with_limit m, %r(String), Type::String
@@ -194,26 +212,38 @@ module ActiveRecord
         register_class_with_limit m, %r(Int128), Type::Integer
         register_class_with_limit m, %r(Int256), Type::Integer
 
-        register_class_with_limit m, %r(Uint8), Type::UnsignedInteger
+        register_class_with_limit m, %r(UInt8), Type::UnsignedInteger
         register_class_with_limit m, %r(UInt16), Type::UnsignedInteger
         register_class_with_limit m, %r(UInt32), Type::UnsignedInteger
         register_class_with_limit m, %r(UInt64), Type::UnsignedInteger
         #register_class_with_limit m, %r(UInt128), Type::UnsignedInteger #not implemnted in clickhouse
         register_class_with_limit m, %r(UInt256), Type::UnsignedInteger
         register_class_with_limit m, %r(Array), Type::String
+        # register_class_with_limit m, %r(Array), Clickhouse::OID::Array
+        # m.register_type(%r(Array)) do |sql_type|
+        #   Clickhouse::OID::Array.new(sql_type)
+        # end
       end
 
       # Quoting time without microseconds
       def quoted_date(value)
         if value.acts_like?(:time)
-          zone_conversion_method = ActiveRecord::Base.default_timezone == :utc ? :getutc : :getlocal
+          if ActiveRecord::version >= Gem::Version.new('7')
+            zone_conversion_method = ActiveRecord.default_timezone == :utc ? :getutc : :getlocal
+          else
+            zone_conversion_method = ActiveRecord::Base.default_timezone == :utc ? :getutc : :getlocal
+          end
 
           if value.respond_to?(zone_conversion_method)
             value = value.send(zone_conversion_method)
           end
         end
 
-        value.to_s(:db)
+        if ActiveRecord::version >= Gem::Version.new('7')
+          value.to_fs(:db)
+        else
+          value.to_s(:db)
+        end
       end
 
       def column_name_for_operation(operation, node) # :nodoc:
@@ -258,26 +288,36 @@ module ActiveRecord
       def create_view(table_name, **options)
         options.merge!(view: true)
         options = apply_replica(table_name, options)
-        td = create_table_definition(apply_cluster(table_name), options)
+        td = create_table_definition(apply_cluster(table_name), **options)
         yield td if block_given?
 
         if options[:force]
           drop_table(table_name, options.merge(if_exists: true))
         end
 
-        execute schema_creation.accept td
+        do_execute(schema_creation.accept(td), format: nil)
       end
 
-      def create_table(table_name, **options)
+      def create_table(table_name, **options, &block)
         options = apply_replica(table_name, options)
-        td = create_table_definition(apply_cluster(table_name), options)
-        yield td if block_given?
+        td = create_table_definition(apply_cluster(table_name), **options)
+        block.call td if block_given?
 
         if options[:force]
           drop_table(table_name, options.merge(if_exists: true))
         end
 
-        execute schema_creation.accept td
+        do_execute(schema_creation.accept(td), format: nil)
+
+        if options[:with_distributed]
+          distributed_table_name = options.delete(:with_distributed)
+          sharding_key = options.delete(:sharding_key) || 'rand()'
+          raise 'Set a cluster' unless cluster
+
+          distributed_options =
+            "Distributed(#{cluster}, #{@config[:database]}, #{table_name}, #{sharding_key})"
+          create_table(distributed_table_name, **options.merge(options: distributed_options), &block)
+        end
       end
 
       # Drops a ClickHouse database.
@@ -295,6 +335,11 @@ module ActiveRecord
 
       def drop_table(table_name, options = {}) # :nodoc:
         do_execute apply_cluster "DROP TABLE#{' IF EXISTS' if options[:if_exists]} #{quote_table_name(table_name)}"
+
+        if options[:with_distributed]
+          distributed_table_name = options.delete(:with_distributed)
+          drop_table(distributed_table_name, **options)
+        end
       end
 
       def change_column(table_name, column_name, type, options = {})
@@ -320,12 +365,45 @@ module ActiveRecord
         @full_config[:replica_name]
       end
 
+      def use_default_replicated_merge_tree_params?
+        database_engine_atomic? && @full_config[:use_default_replicated_merge_tree_params]
+      end
+
+      def use_replica?
+        (replica || use_default_replicated_merge_tree_params?) && cluster
+      end
+
       def replica_path(table)
         "/clickhouse/tables/#{cluster}/#{@config[:database]}.#{table}"
       end
 
+      def database_engine_atomic?
+        current_database_engine = "select engine from system.databases where name = '#{@config[:database]}'"
+        res = select_one(current_database_engine)
+        res['engine'] == 'Atomic' if res
+      end
+
       def apply_cluster(sql)
-        cluster ? "#{sql} ON CLUSTER #{cluster}" : sql
+        if cluster
+          normalized_cluster_name = cluster.start_with?('{') ? "'#{cluster}'" : cluster
+
+          "#{sql} ON CLUSTER #{normalized_cluster_name}"
+        else
+          sql
+        end
+      end
+
+      def supports_insert_on_duplicate_skip?
+        true
+      end
+
+      def supports_insert_on_duplicate_update?
+        true
+      end
+
+      def build_insert_sql(insert) # :nodoc:
+        sql = +"INSERT #{insert.into} #{insert.values_list}"
+        sql
       end
 
       def supports_insert_on_duplicate_skip?
@@ -362,17 +440,32 @@ module ActiveRecord
         @connection.read_timeout = @connection_parameters[:read_timeout] if @connection_parameters[:read_timeout]
         @connection.write_timeout = @connection_parameters[:write_timeout] if @connection_parameters[:write_timeout]
 
+        # Use clickhouse default keep_alive_timeout value of 10, rather than Net::HTTP's default of 2
+        @connection.keep_alive_timeout = @connection_parameters[:keep_alive_timeout] || 10
+
         @connection
       end
 
       def apply_replica(table, options)
-        if replica && cluster && options[:options]
-          match = options[:options].match(/^(.*?MergeTree)\(([^\)]*)\)(.*?)$/)
-          if match
-            options[:options] = "Replicated#{match[1]}(#{([replica_path(table), replica].map{|v| "'#{v}'"} + [match[2].presence]).compact.join(', ')})#{match[3]}"
+        if use_replica? && options[:options]
+          if options[:options].match(/^Replicated/)
+            raise 'Do not try create Replicated table. It will be configured based on the *MergeTree engine.'
           end
+
+          options[:options] = configure_replica(table, options[:options])
         end
         options
+      end
+
+      def configure_replica(table, options)
+        match = options.match(/^(.*?MergeTree)(?:\(([^\)]*)\))?((?:.|\n)*)/)
+        return options unless match
+
+        if replica
+          engine_params = ([replica_path(table), replica].map { |v| "'#{v}'" } + [match[2].presence]).compact.join(', ')
+        end
+
+        "Replicated#{match[1]}(#{engine_params})#{match[3]}"
       end
     end
   end