+++ spec/spec_helper.rb
@@ -7,6 +7,7 @@ require 'clickhouse-activerecord'
 require 'active_support/testing/stream'
 
 FIXTURES_PATH = File.join(File.dirname(__FILE__), 'fixtures')
+CLUSTER_NAME = 'test'
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
@@ -21,9 +22,13 @@ RSpec.configure do |config|
   end
 
   config.around(:each, :migrations) do |example|
-    ActiveRecord::Base.connection.tables.each { |table| ActiveRecord::Base.connection.execute("DROP TABLE #{table}") }
+    clear_consts
+    clear_db
+
     example.run
-    ActiveRecord::Base.connection.tables.each { |table| ActiveRecord::Base.connection.execute("DROP TABLE #{table}") }
+
+    clear_consts
+    clear_db
   end
 end
 
@@ -44,5 +49,30 @@ def schema(model)
   model.reset_column_information
   model.columns.each_with_object({}) do |c, h|
     h[c.name] = c
+  end
+end
+
+def clear_db
+  cluster = ActiveRecord::Base.connection_db_config.configuration_hash[:cluster_name]
+  pattern = if cluster
+              normalized_cluster_name = cluster.start_with?('{') ? "'#{cluster}'" : cluster
+
+              "DROP TABLE %s ON CLUSTER #{normalized_cluster_name}"
+            else
+              'DROP TABLE %s'
+            end
+
+  ActiveRecord::Base.connection.tables.each { |table| ActiveRecord::Base.connection.execute(pattern % table) }
+end
+
+def clear_consts
+  $LOADED_FEATURES.select { |file| file.include? FIXTURES_PATH }.each do |file|
+    const = File.basename(file)
+                .scan(ActiveRecord::Migration::MigrationFilenameRegexp)[0][1]
+                .camelcase
+                .safe_constantize
+
+    Object.send(:remove_const, const.to_s) if const
+    $LOADED_FEATURES.delete(file)
   end
 end